# ARM assembler performance improvements