@rakitzis 2020-10-02
I would need to look deeper. At the time I wrote this R9 was reserved. What changed to allow a leaf routine to clobber it?

It may only surface problems with concurrency. I can't remember at the moment. So... skeptical of the change without evidence either way at the moment. If I find more information I'll post it here.