@greatroar 2020-10-02
Apparently, R9 was declared free in golang/go@89f185fe8a036b0fabce30b20c480cf1c832bdd7. That's sometime before Go 1.4, which is why I didn't know about it. I doubt anyone uses Go 1.4 anymore, but I could add a build tag if you want to be absolutely safe.