@rakitzis 2020-10-02
No that’s ok. Thanks for looking into this! Yeah, we used early versions of the compiler for some time. If you believe R9 is safe to clobber now I am ok with this.