@dchest 2020-10-02
Cool, thank you everyone! I agree, Go 1.4 is already too old, so I'm fine with not supporting everything before it. OK to merge?