@greatroar 2020-10-02
> If you believe R9 is safe to clobber now I am ok with this.

If "you" means me, I'm convinced it's safe. The compiler uses it as a temporary without any saving, too.