@dchest 2020-10-02
Thank you! Merged and tagged as v1.2.2